requirements.txt
asgiref==3.8.1
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
drf-yasg==1.21.7
Pillow==11.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.0